modify: visualization part

+++ scheduler/GB_with_weva.py
@@ -18,7 +18,7 @@ class GB_with_Weva(SchedulerBase):
         self.gamma = 0.2
         self.cls_lr = 0.01
 
-        self.e_drop = self.all_epoch * 0.8 # 30의 0.8
+        self.e_drop = int(self.all_epoch * 0.8) # 30의 0.8
         self.e_end = self.all_epoch
         self.mlast = 3
